Refactor endpoints/integration_test.go
